Split datatypes into complex and primitive
Improve handling of complex and primitive datatypes.

We now have primitive/complex classes and use a variant to union the two.

Added initial support for all the built-in Complex datatypes, but there is only
limited support for actually threading the values through the entire system.
That will come with time.