Emit Complex datatypes as nested classes
We're now splitting out complex dtypes and emitting them as trailing values at
the end of the Swift file.

This modifies the way we handle Complex Types, they're now name/field records,
as opposed to name/component record which datatypes.
This makes things a bit more flexible and recursively handy.